% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc.R
\name{empty_droplets}
\alias{empty_droplets}
\title{Detect Empty Droplets in Single-Cell Experiment Data}
\usage{
empty_droplets(sce_raw, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{sce_raw}{A \code{SingleCellExperiment} object containing raw count data.}

\item{BPPARAM}{A \code{BiocParallelParam} object specifying the parallelization parameters for computation.}

\item{empty_lower_droplets}{An integer specifying the lower bound for the total UMI count to consider a droplet as potentially empty.}
}
\value{
A \code{DataFrame} object with the results of the empty droplet detection, including p-values and other statistics.
}
\description{
This function identifies empty droplets in single-cell experiment data using the \code{DropletUtils::emptyDrops} method.
}
\details{
This function applies the \code{emptyDrops} method from the \code{DropletUtils} package to identify droplets that are likely to be empty based on their UMI counts. The \code{lower} parameter sets the threshold for the minimum UMI count to consider a droplet for testing.
}
\examples{
# Example usage:
library(SingleCellExperiment)
library(DropletUtils)
library(BiocParallel)

# Create a mock SingleCellExperiment object
sce_raw <- SingleCellExperiment(assays = list(counts = matrix(rpois(100, lambda = 1), ncol = 10)))

# Detect empty droplets
result <- tar_sc_single_qc_step_detect_empty_droplets(
  sce_raw = sce_raw,
  empty_lower_droplets = 100,
  BPPARAM = SerialParam()
)

}
