% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_io.R
\name{read10x}
\alias{read10x}
\title{Read 10X Genomics Data}
\usage{
read10x(path)
}
\arguments{
\item{path}{Character. The file path to the 10X Genomics data directory.}
}
\value{
None. (Function implementation is currently incomplete.)
}
\description{
This function reads 10X Genomics single-cell RNA-seq data from the specified path.
}
\examples{
\dontrun{
read10x("/path/to/10x/data")
}

}
