% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc.R
\name{remove_empty_drop}
\alias{remove_empty_drop}
\title{Remove Empty Droplets Based on FDR Threshold}
\usage{
remove_empty_drop(sce, empty_droplets, empty_droplets_fdr_threshold = 0.01)
}
\arguments{
\item{empty_droplets}{A data frame containing information about droplets,
including their FDR values. Can be \code{NULL} if no empty droplet information is available.}

\item{empty_droplets_fdr_threshold}{A numeric value specifying the FDR threshold
for identifying cells. Droplets with FDR values less than or equal to this
threshold are retained.}

\item{sce_raw}{A \code{SingleCellExperiment} object containing raw single-cell data.
Must have non-zero columns.}
}
\value{
A \code{SingleCellExperiment} object with empty droplets removed and an
additional \code{is_empty_fdr} column indicating the FDR values for the retained cells.
}
\description{
This function filters out empty droplets from a SingleCellExperiment (SCE) object
based on a specified false discovery rate (FDR) threshold. It also adds an
\code{is_empty_fdr} column to the SCE object to indicate the FDR values for the retained cells.
}
\details{
\itemize{
\item If \code{empty_droplets} is \code{NULL}, the function adds an \code{is_empty_fdr} column
with \code{NA} values to the SCE object.
\item If \code{empty_droplets} is provided, the function filters the SCE object to
retain only the cells with FDR values less than or equal to the specified threshold.
\item The function ensures that the input \code{sce_raw} has at least one column.
}
}
\examples{
# Example usage:
sce_filtered <- tar_sc_single_qc_step_remove_empty_drop(
  sce_raw = sce_object,
  empty_droplets = droplet_data,
  empty_droplets_fdr_threshold = 0.01
)

}
