% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctargets_single_qc.R
\name{scTarget_single_qc}
\alias{scTarget_single_qc}
\title{Read 10x Genomics Count Matrix}
\usage{
scTarget_single_qc(
  name,
  x10_file_path,
  empty_droplets_fdr_threshold = 0.01,
  empty_lower_droplets = TRUE,
  replace_unfiltered = TRUE,
  save_dataset_sensitive_filtering = TRUE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{path}{A string specifying the path to the 10x Genomics data directory.
This directory should contain the matrix, features, and barcodes files.}
}
\value{
A \code{SingleCellExperiment} object containing the count matrix
and associated metadata.
}
\description{
This function reads a 10x Genomics count matrix from the specified path and
returns a SingleCellExperiment object. The column names of the resulting
object are set to the barcodes from the cell metadata.
}
\examples{
\dontrun{
  # Example usage:
  sce <- tar_sc_single_qc_step_read_10x_counts("path/to/10x/data")
}

}
