% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc_steps.R
\name{tar_create_ready_sce}
\alias{tar_create_ready_sce}
\title{Create a ready SingleCellExperiment (SCE) object from a CSV file}
\usage{
tar_create_ready_sce(file)
}
\arguments{
\item{file}{Character. Path to the input CSV file.}
}
\value{
Character. The filename of the output CSV ("output.csv").
}
\description{
This function reads a CSV file, selects the first few rows of the data,
writes the result to a new CSV file named "output.csv", and returns the
filename.
}
\examples{
tar_create_ready_sce("input.csv")

}
