% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc_steps.R
\name{tar_sc_single_qc_step_after_filter_rowsums}
\alias{tar_sc_single_qc_step_after_filter_rowsums}
\title{Perform Quality Control Step After Filtering Rowsums}
\usage{
tar_sc_single_qc_step_after_filter_rowsums(sce_filter)
}
\arguments{
\item{sce_filter}{A \code{SingleCellExperiment} object that has been filtered.}
}
\value{
A numeric vector containing the row sums of the counts matrix.
}
\description{
This function calculates the row sums of the counts matrix from a
SingleCellExperiment object after filtering.
}
\examples{
# Assuming `sce_filtered` is a filtered SingleCellExperiment object:
row_sums <- tar_sc_single_qc_step_after_filter_rowsums(sce_filtered)

}
