% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc_steps.R
\name{tar_sc_single_qc_step_cal_per_cell_qc_metrics}
\alias{tar_sc_single_qc_step_cal_per_cell_qc_metrics}
\title{Calculate Per-Cell QC Metrics}
\usage{
tar_sc_single_qc_step_cal_per_cell_qc_metrics(sce_no_empty_drop, BPPARAM)
}
\arguments{
\item{sce_no_empty_drop}{A \code{SingleCellExperiment} object that has been filtered to remove empty droplets.}

\item{BPPARAM}{A \code{BiocParallelParam} object specifying the parallelization backend to use.}
}
\value{
A \code{DataFrame} containing per-cell QC metrics, including metrics for mitochondrial and ribosomal subsets.
}
\description{
This function calculates per-cell quality control (QC) metrics for a SingleCellExperiment object.
It identifies mitochondrial and ribosomal genes based on their symbols and computes QC metrics
using the \code{scater::perCellQCMetrics} function.
}
\details{
\itemize{
\item Mitochondrial genes are identified using a case-insensitive regex match for "MT-".
\item Ribosomal genes are identified using a case-insensitive regex match for "^RP\link{SL}".
\item The \code{BPPARAM} parameter allows for parallel computation of QC metrics.
}
}
