% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc_steps.R
\name{tar_sc_single_qc_step_create_unfiltered_sce}
\alias{tar_sc_single_qc_step_create_unfiltered_sce}
\title{Create an Unfiltered SingleCellExperiment Object with QC Metrics}
\usage{
tar_sc_single_qc_step_create_unfiltered_sce(
  sce_no_empty_drop,
  per_cell_qc_metrics,
  replace
)
}
\arguments{
\item{sce_no_empty_drop}{A \code{SingleCellExperiment} object containing the input data
without empty droplets.}

\item{per_cell_qc_metrics}{A \code{data.frame} containing per-cell quality control metrics.
Expected columns include \code{total}, \code{detected}, and \code{subsets_mito_percent}.}

\item{replace}{A logical value indicating whether to replace existing metadata
columns in the SCE object if they overlap with the new QC metrics. Default is \code{FALSE}.}
}
\value{
A \code{SingleCellExperiment} object with updated metadata columns containing
the QC filtering results.
}
\description{
This function applies dataset-sensitive and custom cell filtering criteria
to a SingleCellExperiment (SCE) object and appends the filtering results
as metadata columns. The function is designed to handle quality control (QC)
metrics and allows for the replacement of existing metadata columns if specified.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Dataset-sensitive cell filtering}: Identifies outliers based on library size,
number of detected features, and mitochondrial content using the Median Absolute
Deviation (MAD) method.
\item \strong{Custom cell filtering}: Applies user-defined thresholds for library size,
number of detected features, and mitochondrial content.
\item Combines the results of the dataset-sensitive and custom filters using logical
AND operations.
\item Appends the filtering results as new metadata columns (\code{discard_qc} and
\code{discard_custom}) to the SCE object.
}
}
\examples{
# Example usage:
sce <- tar_sc_single_qc_step_create_unfiltered_sce(
  sce_no_empty_drop = sce_object,
  per_cell_qc_metrics = qc_metrics,
  replace = TRUE
)

}
