% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sc_single_qc_steps.R
\name{tar_sc_single_step_make_custom_filter}
\alias{tar_sc_single_step_make_custom_filter}
\title{Apply Custom Filtering to SingleCellExperiment Object}
\usage{
tar_sc_single_step_make_custom_filter(sce_unfiltered)
}
\arguments{
\item{sce_unfiltered}{A SingleCellExperiment object containing unfiltered data.
The object must have a logical column \code{discard_custom} in its colData,
where \code{TRUE} indicates cells to be discarded.}
}
\value{
A SingleCellExperiment object with cells that are not marked
for discarding based on the \code{discard_custom} column.
}
\description{
This function filters a SingleCellExperiment object by removing cells
that are marked for discarding based on a custom filter.
}
\examples{
# Assuming `sce` is a SingleCellExperiment object with a `discard_custom` column:
filtered_sce <- tar_sc_single_step_make_custom_filter(sce)

}
